really ugly thing: workarround for snmp package
For now, if we have installed *snmp* on the server, before nagios-plugins-all,
this one cannot be installed...
So we have to ensure this package is installed before any snmp package.
The other solution is to install each nagios-plugins-* one by one, omitting the snmp one,
but this won't fixe the problem.